etl_ontology/requirements.txt
beautifulsoup4
dateparser
numpy
odfpy
openpyxl
pandas
PyYAML
python-dateutil
rdflib
requests
tqdm
datetime